@@ -1,10 +1,14 @@
 {
   "name": "react-highlighter",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "description": "Highlight select fragments of texts",
   "main": "lib/highlighter.js",
+  "files": [
+    "lib",
+    "dist"
+  ],
   "peerDependencies": {
-    "react": "^0.14.0"
+    "react": "^0.14.0 || ^15.0.0"
   },
   "scripts": {
     "test": "_mocha",
@@ -29,16 +33,19 @@
   },
   "homepage": "https://github.com/helior/react-highlighter",
   "devDependencies": {
-    "chai": "^2.1.2",
-    "coveralls": "^2.11.2",
-    "istanbul": "^0.4.1",
-    "jsdom": "^3.1.2",
-    "mocha": "^2.3.4",
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0",
-    "webpack": "^1.7.3"
+    "chai": "^3.5.0",
+    "coveralls": "^2.11.9",
+    "istanbul": "^0.4.4",
+    "jsdom": "^9.4.0",
+    "mocha": "^2.5.3",
+    "react": "^15.2.0",
+    "react-dom": "^15.2.0",
+    "webpack": "^1.13.1"
   },
   "dependencies": {
-    "escape-string-regexp": "^1.0.4"
+    "blacklist": "^1.1.2",
+    "create-react-class": "^15.5.2",
+    "escape-string-regexp": "^1.0.5",
+    "prop-types": "^15.5.8"
   }
 }
