@@ -1,22 +1,24 @@
 var React = require('react');
 var RegExpPropType = require('./regExpPropType');
 var escapeStringRegexp = require('escape-string-regexp');
+var blacklist = require('blacklist');
+var createReactClass = require('create-react-class');
+var PropTypes = require('prop-types');
 
-
-var Highlighter = React.createClass({displayName: "Highlighter",
+var Highlighter = createReactClass({displayName: "Highlighter",
   count: 0,
 
   propTypes: {
-    search: React.PropTypes.oneOfType([
-      React.PropTypes.string,
-      React.PropTypes.number,
-      React.PropTypes.bool,
+    search: PropTypes.oneOfType([
+      PropTypes.string,
+      PropTypes.number,
+      PropTypes.bool,
       RegExpPropType
     ]).isRequired,
-    caseSensitive: React.PropTypes.bool,
-    matchElement: React.PropTypes.string,
-    matchClass: React.PropTypes.string,
-    matchStyle: React.PropTypes.object
+    caseSensitive: PropTypes.bool,
+    matchElement: PropTypes.string,
+    matchClass: PropTypes.string,
+    matchStyle: PropTypes.object
   },
 
   getDefaultProps: function() {
@@ -29,7 +31,9 @@ var Highlighter = React.createClass({displayName: "Highlighter",
   },
 
   render: function() {
-    return React.DOM.span(React.__spread({}, this.props), this.renderElement(this.props.children));
+    var props = blacklist(this.props, 'search', 'caseSensitive', 'matchElement', 'matchClass', 'matchStyle');
+
+    return React.createElement('span', props, this.renderElement(this.props.children));
   },
 
   /**
@@ -170,7 +174,7 @@ var Highlighter = React.createClass({displayName: "Highlighter",
    */
   renderPlain: function(string) {
     this.count++;
-    return React.DOM.span({'key': this.count}, string);
+    return React.DOM.span({'key': string}, string);
   },
 
   /**
@@ -184,7 +188,7 @@ var Highlighter = React.createClass({displayName: "Highlighter",
   renderHighlight: function(string) {
     this.count++;
     return React.DOM[this.props.matchElement]({
-      key: this.count,
+      key: string,
       className: this.props.matchClass,
       style: this.props.matchStyle
     }, string);
